# api/src/main/java/com/zcommerce/platform/controller/CategoriesApiImpl.java
package com.zcommerce.platform.controller;

import com.zcommerce.platform.api.CategoriesApi;
import com.zcommerce.platform.api.model.*;
import com.zcommerce.platform.domain.entity.Category;
import com.zcommerce.platform.domain.repository.CategoryRepository;
import java.util.List;
import java.util.stream.Collectors;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

@RestController
@Slf4j
public class CategoriesApiImpl implements CategoriesApi {

  private final CategoryRepository categoryRepository;

  public CategoriesApiImpl(CategoryRepository categoryRepository) {
    this.categoryRepository = categoryRepository;
  }

  @Override
  public ResponseEntity<List<CategoryResponseApi>> getAllCategories() {
    log.info("Getting all categories");
    List<Category> categories = categoryRepository.findAll();
    return ResponseEntity.ok(
        categories.stream().map(this::convertToApiModel).collect(Collectors.toList()));
  }

  @Override
  public ResponseEntity<CategoryResponseApi> getCategoryById(Long categoryId) {
    log.info("Getting category by ID: {}", categoryId);
    return categoryRepository
        .findById(categoryId)
        .map(category -> ResponseEntity.ok(convertToApiModel(category)))
        .orElse(ResponseEntity.status(404).body(null));
  }

  private CategoryResponseApi convertToApiModel(Category entity) {
    var apiModel = new CategoryResponseApi();
    apiModel.setId(entity.getId());
    apiModel.setName(entity.getName());
    apiModel.setDescription(entity.getDescription());
    apiModel.setIsActive(entity.getActive());
    return apiModel;
  }
}