# api/src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: Z-Commerce Platform API
  version: 1.0.0
  description: Z-Commerce e-commerce platform API for product and order management
servers:
  - url: http://localhost:8080/zcommerce
    description: Local development
  - url: http://13.202.157.59:8080/zcommerce
    description: AWS EC2
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
        - firstName
        - lastName
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
...
    CustomerResponseApi:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        address:
          $ref: '#/components/schemas/AddressResponseApi'
        role:
          type: string
        isActive:
          type: boolean
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        orderCount:
          type: integer
          format: int64
    AddressResponseApi:
      type: object
      properties:
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    UpdateCustomerRequestApi:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        streetAddress:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    CustomerStatistics:
      type: object
      properties:
        totalCustomers:
          type: integer
          format: int64
        activeCustomers:
          type: integer
          format: int64
        verifiedEmails:
          type: integer
          format: int64

    # Product
...
    CategoryResponseApi:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean

    # Order
...
paths:
...
  /api/v1/customers:
    post:
      tags: [Customers]
      operationId: createCustomer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'

    get:
      tags: [Customers]
      operationId: getAllCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/customers/{id}:
    get:
      tags: [Customers]
      operationId: getCustomerById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'

    put:
      tags: [Customers]
      operationId: updateCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequestApi'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'

    delete:
      tags: [Customers]
      operationId: deleteCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/customers/username/{username}:
    get:
      tags: [Customers]
      operationId: getCustomerByUsername
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'
  /api/v1/customers/{id}/activate:
    patch:
      tags: [Customers]
      operationId: activateCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'
  /api/v1/customers/{id}/deactivate:
    patch:
      tags: [Customers]
      operationId: deactivateCustomer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'
  /api/v1/customers/{id}/verify-email:
    patch:
      tags: [Customers]
      operationId: verifyCustomerEmail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseApi'
  /api/v1/customers/search:
    get:
      tags: [Customers]
      operationId: searchCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/customers/role/{role}:
    get:
      tags: [Customers]
      operationId: getCustomersByRole
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/customers/active:
    get:
      tags: [Customers]
      operationId: getActiveCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/customers/city/{city}:
    get:
      tags: [Customers]
      operationId: getCustomersByCity
      security:
        - bearerAuth: []
      parameters:
        - name: city
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
  /api/v1/customers/top:
    get:
      tags: [Customers]
      operationId: getTopCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponseApi'
  /api/v1/customers/statistics:
    get:
      tags: [Customers]
      operationId: getCustomerStatistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerStatistics'

  # Payments
...
  /api/v1/categories/{categoryId}:
    get:
      tags: [Categories]
      operationId: getCategoryById
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseApi'